Step 4 - Creating an Identity Microservice
Added an Identity Microsoft and installed OAuth and OpenIDConnect to handle logging in to the API. The basic setup includes Login and Logout pages, but I had to create Register page from scratch. Added a custom profile to the service so the user and username are included in the token on the resource server. Next commit will focus on adding Gateway to our project.