Step 5 - Adding a Gateway Service
Added a gateway service (Reverse Proxy) to allow the individual microservices to communicate directly with the client-side app (website) as well as an indirect link to each other. Also to prepare for dockerization in the next commit, I added the remaining consumers to both the Auction and Search microservices. Also added a new client to the Identity microservice to return data to the website and not just Postman.